+++ module/settings.js
@@ -337,47 +337,11 @@ export default function registerSystemSettings() {
     },
   });
 
-  // obsolete - will be removed at some point
-  game.settings.register("deltagreen", "characterSheetFont", {
-    name: "World Font Choice",
-    hint: "Choose font style for use throughout this world.",
-    scope: "world", // This specifies a world-stored setting
-    config: false, // This specifies that the setting appears in the configuration view
-    type: String,
-    choices: {
-      // If choices are defined, the resulting setting will be a select menu
-      SpecialElite: "Special Elite (Classic Typewriters Font)",
-      Martel: "Martel (Clean Modern Font)",
-      Signika: "Signika (Foundry Default Font)",
-      TypeWriterCondensed:
-        "Condensed Typewriter (Modern, Small Typewriter Font)",
-      PublicSans: "Public Sans (US Government-style sans serif font)",
-      // "atwriter": "Another Typewriter (Alternate Old-style Typewriter Font)"
-    },
-    default: "SpecialElite", // The default value for the setting
-    onChange: (value) => {
-      // A callback function which triggers when the setting is changed
-      // console.log(value)
-    },
-  });
-
-  // obsolete - will be removed at some point
-  game.settings.register("deltagreen", "characterSheetBackgroundImageSetting", {
-    name: "World Sheet Background Image",
-    hint: "Choose background image for use throughout this world. (Refresh page to see change.)",
-    scope: "world", // This specifies a world-stored setting
-    config: false, // This specifies that the setting appears in the configuration view
-    type: String,
-    choices: {
-      // If choices are defined, the resulting setting will be a select menu
-      OldPaper1: "Old Dirty Paper (Good with Special Elite Font)",
-      IvoryPaper: "Ivory White Paper (Good with Martel Font)",
-      DefaultParchment: "Default Parchment (Good with Signika Font)",
-    },
-    default: "OldPaper1", // The default value for the setting
-    onChange: (value) => {
-      // A callback function which triggers when the setting is changed
-      // console.log(value)
-    },
+  game.settings.register(DG.ID, "schemaMigrationVersion", {
+    name: "Schema Migration Version",
+    scope: "world",
+    config: false,
+    type: Number,
+    default: 0,
   });
 }